[FE-214]feat: Card컴포넌트화 -중복 사용됨에 따라 Card컴포넌트화 -TogetherSlider에서 Card컴포넌트로 대체
